move util into base/auxiliary
