+++ techsupport_bot/extensions/duck.py
@@ -8,7 +8,6 @@ from datetime import timedelta
 import base
 import discord
 import ui
-import util
 from base import auxiliary
 from discord import Color as embed_colors
 from discord.ext import commands
@@ -343,7 +342,7 @@ class DuckHunt(base.LoopCog):
         # Executed if there are no/invalid args supplied
         await base.extension_help(self, ctx, self.__module__[11:])
 
-    @util.with_typing
+    @auxiliary.with_typing
     @commands.guild_only()
     @duck.command(
         brief="Get duck stats",
@@ -382,7 +381,7 @@ class DuckHunt(base.LoopCog):
 
         await ctx.send(embed=embed)
 
-    @util.with_typing
+    @auxiliary.with_typing
     @commands.guild_only()
     @duck.command(
         brief="Get duck friendship scores",
@@ -437,7 +436,7 @@ class DuckHunt(base.LoopCog):
 
         await ui.PaginateView().send(ctx.channel, ctx.author, embeds)
 
-    @util.with_typing
+    @auxiliary.with_typing
     @commands.guild_only()
     @duck.command(
         brief="Get the record holder",
@@ -472,7 +471,7 @@ class DuckHunt(base.LoopCog):
 
         await ctx.send(embed=embed)
 
-    @util.with_typing
+    @auxiliary.with_typing
     @commands.guild_only()
     @duck.command(
         brief="Get duck kill scores",
@@ -535,7 +534,7 @@ class DuckHunt(base.LoopCog):
             user_text_extra = ""
         return f"{user_text}{user_text_extra}"
 
-    @util.with_typing
+    @auxiliary.with_typing
     @commands.guild_only()
     @duck.command(
         brief="Releases a duck into the wild",
@@ -564,7 +563,7 @@ class DuckHunt(base.LoopCog):
             channel=ctx.channel,
         )
 
-    @util.with_typing
+    @auxiliary.with_typing
     @commands.cooldown(1, 600)
     @commands.guild_only()
     @duck.command(
@@ -613,7 +612,7 @@ class DuckHunt(base.LoopCog):
             channel=ctx.channel,
         )
 
-    @util.with_typing
+    @auxiliary.with_typing
     @commands.cooldown(1, 600)
     @commands.guild_only()
     @duck.command(
@@ -679,7 +678,7 @@ class DuckHunt(base.LoopCog):
             channel=ctx.channel,
         )
 
-    @util.with_typing
+    @auxiliary.with_typing
     @commands.has_permissions(administrator=True)
     @commands.guild_only()
     @duck.command(
