+++ techsupport_bot/extensions/grab.py
@@ -7,7 +7,6 @@ import random
 import base
 import discord
 import ui
-import util
 from base import auxiliary
 from discord.ext import commands
 
@@ -67,7 +66,7 @@ class Grabber(base.BaseCog):
     HAS_CONFIG = False
     SEARCH_LIMIT = 20
 
-    @util.with_typing
+    @auxiliary.with_typing
     @commands.guild_only()
     @commands.check(invalid_channel)
     @commands.command(
@@ -146,7 +145,7 @@ class Grabber(base.BaseCog):
         # Executed if there are no/invalid args supplied
         await base.extension_help(self, ctx, self.__module__[11:])
 
-    @util.with_typing
+    @auxiliary.with_typing
     @commands.guild_only()
     @commands.check(invalid_channel)
     @grabs.command(
@@ -216,7 +215,7 @@ class Grabber(base.BaseCog):
 
         await ui.PaginateView().send(ctx.channel, ctx.author, embeds)
 
-    @util.with_typing
+    @auxiliary.with_typing
     @commands.guild_only()
     @commands.check(invalid_channel)
     @grabs.command(
@@ -272,7 +271,7 @@ class Grabber(base.BaseCog):
 
         await ctx.send(embed=embed)
 
-    @util.with_typing
+    @auxiliary.with_typing
     @commands.guild_only()
     @commands.check(invalid_channel)
     @grabs.command(
