+++ techsupport_bot/extensions/poll.py
@@ -6,7 +6,6 @@ import json
 import base
 import discord
 import emoji
-import util
 from base import auxiliary
 from discord.ext import commands
 
@@ -111,7 +110,7 @@ class ReactionPoller(PollGenerator):
         # Executed if there are no/invalid args supplied
         await base.extension_help(self, ctx, self.__module__[11:])
 
-    @util.with_typing
+    @auxiliary.with_typing
     @poll.command(
         brief="Shows example poll JSON",
         description="Shows what JSON to upload to generate a poll",
@@ -124,7 +123,7 @@ class ReactionPoller(PollGenerator):
         )
         await ctx.send(file=json_file)
 
-    @util.with_typing
+    @auxiliary.with_typing
     @commands.guild_only()
     @poll.command(
         aliases=["create"],
@@ -136,7 +135,7 @@ class ReactionPoller(PollGenerator):
     )
     async def generate(self, ctx):
         """Method to generate the poll for discord."""
-        request_body = await util.get_json_from_attachments(ctx.message)
+        request_body = await auxiliary.get_json_from_attachments(ctx.message)
         if not request_body:
             await auxiliary.send_deny_embed(
                 message="I couldn't find any data in your upload", channel=ctx.channel
@@ -275,7 +274,7 @@ class StrawPoller(PollGenerator):
         # Executed if there are no/invalid args supplied
         await base.extension_help(self, ctx, self.__module__[11:])
 
-    @util.with_typing
+    @auxiliary.with_typing
     @strawpoll.command(
         brief="Shows example poll JSON",
         description="Shows what JSON to upload to generate a poll",
@@ -288,7 +287,7 @@ class StrawPoller(PollGenerator):
         )
         await ctx.send(file=json_file)
 
-    @util.with_typing
+    @auxiliary.with_typing
     @strawpoll.command(
         brief="Generates a strawpoll",
         description="Returns a link to a Strawpoll generated by args",
@@ -296,7 +295,7 @@ class StrawPoller(PollGenerator):
     )
     async def generate(self, ctx):
         """Method to generate the poll form the discord command."""
-        request_body = await util.get_json_from_attachments(ctx.message)
+        request_body = await auxiliary.get_json_from_attachments(ctx.message)
         if not request_body:
             await auxiliary.send_deny_embed(
                 message="I couldn't find any data in your upload", channel=ctx.channel
